Fix setting and unsetting pager environment variable
Always enable the pager when setting pager=1 and disable it otherwise,
rather than incrementing and decrementing an internal variable.  The
existing behavior required two "set pager=0" to turn off after two "set
pager=1", and effectively enabled the pager when running an unmatched
"set pager=0" (since -1 is non-zero) and disabled it with a subsequent
"set pager=1".