Fix couple of linker complaints when built with the 2.34 version of linker
a) Fix the linker warning "-r and -pie may not be used together"

The fix has been ported from the Qemu commmit:
c96f0ee6a67ca ("rules.mak: Use -r instead of -Wl, -r to fix building when PIE is default")

With PIE enabled by deafault, we should use "-r" instead of "Wl,-r". The gcc
"-r" option avoids gcc from sending "-pie" to the linker when a relocatable
object is sent as its input.

b) Fix linker warning "PHDR segment not covered by LOAD segment".

To fix this, we send "-static" as the linker argument so that PHDR segment
is not generated.

See also the following note in ld:

Changes in 2.34:

* The ld check for "PHDR segment not covered by LOAD segment" is more
  effective, catching cases that were wrongly allowed by previous versions of
  ld.  If you see this error it is likely you are linking with a bad linker
  script or the binary you are building is not intended to be loaded by a
  dynamic loader.  In the latter case --no-dynamic-linker is appropriate.

Signed-off-by: Ani Sinha <ani@anisinha.ca>