grub-mkimage: make it link statically so that it can be used on any host
grub-mkimage is used by grub-mkrescue which in turn is used to generate the
bios bits iso file. We need to ensure that grub-mkimage can run on any host,
even on those that have an earlier version of glibc installed. Hence, we make
it statically linked executable so that the exact version of libraries that are
available on the host where its run does not matter. Thus issues such as the
following can be avoided when executed:

grub-mkimage: /lib/x86_64-linux-gnu/libc.so.6: version `GLIBC_2.33' not found

Signed-off-by: Ani Sinha <ani@anisinha.ca>